requirements.txt
# Extras --------------------------------------
ipython  # interactive notebook
psutil  # system utilization
thop>=0.1.1  # FLOPs computation
streamlit
shutil
base64
ultralytics
Pillow  # Python Imaging Library
# albumentations>=1.0.3
# pycocotools>=2.0  # COCO mAP
# roboflow